Add a LocalCache supporting the W-TinyLFU eviction policy
Window TinyLFU (W-TinyLFU) is a modern eviction policy that
outperforms the currently implemented algorithms: LRU, LFU,
and Hybrid (default).

W-TinyLFU utilizes a popularity sketch, CountMin, to compactly
estimate the frequency over a time period. This sketch is used
as a frequency filter by comparing the entry being added to the
cache to eviction policy's victim. If the victim is more popular
then the candidate is discarded, which avoids polluting the cache
with "one-hit wonders". The sketch is aged by halving all of its
counters periodically. Because it only needs to detect heavy
hitters, a matrix of 4-bit counters is used.

In recency-biased workloads, aggressively filtering by the
frequency can reduce the hit rate. Therefore the cache is split
into two LRUs, one before and one after the filter, and uses
TinyLFU to conditionally promote. Thus the admission window's
viction and the main space's compete. Since the main region is
frequency-biased, it uses SLRU to choose a better victim.

The optimal relative size of these two regions is workload
dependent. It is configured dynamically by using hill climbing.
The hit rate is sampled and the admission window's size is
increased or decreased depending on whether the last guess was
beneficial. Eventually the climber oscillates around the best
configuration, so the step size decays gradually to allow for
convergence. An adaption will restart if the hit rate change
dramatically, indicating new different workload.

The hit rate across a variety of workloads shows a significant
improvement.
 - glimpse: An analytical loop
 - multi3: four concurrent workloads
 - corda: blockchain mining
 - loop/corda/loop: chaining of mru/lru/mru traces
 - oltp: from a CODASYL database
 - ds1: from a SQL database
 - s3: from a web search engine

| Trace | Cache Size | W-TinyLFU | Hybrid | Lru | Optimal |
|---|---|---|---|---|---|
| glimpse | 512 | **32.65%** | 2.64% | 0.98% | 35.13% |
| multi3 | 512 | **44.16%** | 31.41% | 32.78% | 49.10% |
| corda | 512 | 32.29% | 12.14% | **33.33%** | 33.33% |
| loop/corda/loop | 512 | **36.44%** | 14.95% | 21.65% | 39.38% |
| oltp | 1024 | **40.21%** | 15.39% | 33.22% | 53.86% |
| ds1 | 4M | **44.76%** | 36.95% | 20.24% | 48.09% |
| s3 | 400K | **42.52%** | 31.32% | 12.04% | 59.96% |